projects/rt-akhmarov/pyproject.toml
[project]
name = "rt-akhmarov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "hf-transfer>=0.1.9",
    "langchain>=1.2.4",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langfuse[langchain]>=3.11.2",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.16.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
]
